+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+##### 2.1.1:
+    Added `cambi_opt` options - `max_log_contrast`, `enc_bitdepth`, `eotf`.
+    Added function `VMAF2`.
+    Clips must have either 8-bit or 10-bit bit depth.
+
 ##### 2.1.0:
     Added parameter cambi_opt.
     Removed model default value.
@@ -13,6 +18,6 @@
 
 ##### 2.0.0:
     Updated to libvmaf 2.x.
-    
+
 ##### 1.0.0:
     Initial release.